requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1

# Core conflict resolution for mediapipe vs google-cloud
google-cloud-storage==2.9.0
google-cloud-tasks==2.13.0
google-cloud-firestore==2.11.0

# LLM Integration
google-generativeai==0.8.3

# MediaPipe and its dependencies
mediapipe==0.10.14
opencv-python-headless==4.10.0.84
numpy==1.26.4